pyproject.toml
[project]
name = "mammoth"
version = "0.2.2"
readme = "README.md"
description = "The Mammoth framework for Continual Learning research."
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "The Mammoth Team" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: GPU :: NVIDIA CUDA"
]
keywords = [
    "continual-learning",
    "deep-learning",
    "pytorch",
]
dependencies = [
    "accelerate>=1.6.0",
    "bitsandbytes>=0.45.4",
    "clip",
    "datasets>=4.5.0",
    "ftfy>=6.3.1",
    "googledrivedownloader==0.4",
    "gpustat>=1.1.1",
    "kornia>=0.7.0",
    "numpy>=2.2.4",
    "onedrivedownloader>=1.1.3",
    "open-clip-torch>=2.32.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "sentencepiece>=0.2.0",
    "timm==0.9.8",
    "torch>=2.1.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "wandb>=0.20.1",
]

[project.urls]
Repository = "https://github.com/aimagelab/mammoth"
Documentation = "https://aimagelab.github.io/mammoth/"

[project.optional-dependencies]
extra = [
    "accelerate>=1.6.0",
    "bitsandbytes>=0.45.4",
    "clip",
    "decorator>=5.2.1",
    "googledrivedownloader==0.4",
    "kmeans-pytorch>=0.3",
    "onedrivedownloader==1.1.3",
    "pandas>=2.2.3",
    "qpsolvers",
    "scikit-learn>=1.6.1",
    "setproctitle==1.3.2",
    "timm==0.9.8",
    "transformers>=4.50.3",
    "wandb>=0.19.9",
    "xitorch>=0.5.1",
    "dotenv",
    "gdown",
    "matplotlib",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }

[tool.uv.workspace]
members = ["docs"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.16.0",
    "pandas-stubs>=2.2.3.250527",
    "pytest>=8.3.5",
    "types-pyyaml>=6.0.12.20250516",
    "types-six>=1.17.0.20250515",
    "types-tqdm>=4.67.0.20250516",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = "tests"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore::FutureWarning",
    "ignore::Warning",
    "ignore::SyntaxWarning",
]

[tool.ruff.lint]
ignore = [
    "E402"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mammoth"]
py-modules = []
include-package-data = true

[tool.setuptools.package-dir]
mammoth = "."
